envoy: convert err msg to debug msg
We were logging an error when trying to update an ingress that no longer
exists. This is not really an error, it could happen if the ingress was
deleted while running the loop to update the config.

I noticed this error while running the "TestRouteVisibilityChanges" test
of the Knative e2e suite.