Convert to knative controller
This commit reorganizes the code to use a Knative controller.
That changes a few things:
- We no longer need custom watchers for Ingresses and Endpoints.
- The code has been changed to use Ingress instead of IngressAccessors.
The reason is that IngressAccessors have been removed from the Knative
codebase recently. Also, they provide an ingressInformer, but not an
ingressAccessorInformer. This is a change that we'd need to do anyway.

This is a first step. There are at least a couple of things that need
to be done:
- The part that checks that the gateways are in sync, should be moved to
a goroutine so it does not block the controller loop.
- We need to adapt the code to use the endpoint lister in the
reconciler. Also, we'll need to evaluate whether we want to react to
each endpoint event.