generator: propagate err when secret not available
This is needed to pass one of the Knative serving tests:
"TestPerKsvcCert_localCA". The secrets referenced in the TLS section of
an ingress might not be available when the ingress is created. What we
need to do in that case is propagate the error to the reconciler, so the
event is retried later until the secret is created.