generator: extract IngressTranslator
This commits extracts the logic that extract the Envoy structures from
a Knative Ingress.

I think this makes it clearer the Envoy structures that we need to
extract from an Ingress, and also allows us to reduce the number of
parameters that are passed from the reconciler. Now we can just pass an
IngressTranslator instance.